Clear test data before recording results (Jenkins)
Summary:
This is a sledgehammer to the problem of intermediate failures in some areas. Right now we make the assumption that TestResult.save() doesn't have to consider if there's existing data, which allows us to make it much more efficient. However, there are some cases where this assumption ends up incorrect and it winds up with a bunch of duplicate key errors.

Additionally this ups the lock time on sync_build to 60 seconds.

Reviewers: cf, mattr

Reviewed By: mattr

Differential Revision: https://tails.corp.dropbox.com/D31683