Move Jenkins queue/active sync into sync_step
Summary:
This will let us more correctly support real-time syncing from subjobs since we'll be able to fire off jobs to detect them in the queue.

It also greatly reduces the amount of work in the queue as sync_job is a no-op (other than checking if steps have completed).

Reviewers: cf

Differential Revision: https://tails.corp.dropbox.com/D34141