Default parameters to revision values
Summary:
If a vcs backend is set for a repository it attempt to fetch the commit and default the author, label, and message values. If this is a newly discovered commit it also creates the corresponding revision row.

This has the added benefit of allowing us to specify things like 'default' as a sha to the entrypoint.

Reviewers: cf, nipunn, josiah

Reviewed By: cf

Differential Revision: https://tails.corp.dropbox.com/D37515