Add JenkinsCollectorBuildStep
Summary:
The collector step looks for a jobs.json in the collected artifacts, and given that it will fanout
into additional jobs.

This hacks around the current situation that we only support the first/single BuildStep, but in the
future this would ideally create a second build step for the job fanout.

Reviewers: cf

Reviewed By: cf

Differential Revision: https://tails.corp.dropbox.com/D45269