Add file whitelist option for commit triggers
Summary:
The file whitelist causes a "which files were changed" check on the commit trigger handler.

The whitelist is a list of patterns (newline separated) which allow wildcard matches via an asterix.

i.e.

```
metaserver/*
blockserver/*
```

The setting is managed via project settings, and is stored as build.file-whitelist in project options.

Reviewers: vishal

Reviewed By: vishal

Subscribers: rajat, ar, mattr, wwu, haoyi, cristina

Differential Revision: https://tails.corp.dropbox.com/D69134