[changes] Specify cluster too in the config
Summary:
Address kyle's comment re: cluster.

I had verified that cluster is not used for selecting the node and so assumed that we need not pass it. But after the request is sent, jenkins tried to find a machine with the given CLUSTER tag. So we need to pass the right cluster info

Test Plan: unit tests

Reviewers: kylec, vishal

Reviewed By: vishal

Subscribers: wwu

Differential Revision: https://tails.corp.dropbox.com/D84153