Allow any jobstep that isn't finished to timeout
Summary:
We have a number of stuck jobsteps (and we will continue to get them) from Mesos
where the state is not `in_progress` (`allocated' is a primary culprit) and
the jobstep hasn't started.
We want all unfinished jobsteps to be able to timeout; this change makes that
happen.
Since timing out a jobstep that isn't `in_progress` is almost certainly an
infrastructural issue, a warning is sent to Sentry in those cases.

Test Plan: Light unit; suggestions welcome.

Reviewers: josiah

Reviewed By: josiah

Subscribers: changesbot, wwu

Differential Revision: https://tails.corp.dropbox.com/D93802