Remove UnableToGetLock alias, test that jobstep_allocate returns a 503 when unable to get lock
Summary:
Inability to allocate is expected in some circumstances, which is why the code intended to return
a 503 in those cases. The intended exception wasn't caught, though it looks like it should've
been. This change removes the indirection, making it do what it looks to be doing, and verifies
that a 503 is returned in that case.
This should resolve a bunch of errors from uncaught exceptions we see at deploy time.

Test Plan: Unit

Reviewers: jukka

Reviewed By: jukka

Subscribers: changesbot, wwu

Differential Revision: https://tails.corp.dropbox.com/D94591