Render step 'data' JSON with indenting and sorted keys
Summary:
The step configuration data is a config, rendered to JSON for human reading/editing.
Recognizing this, seems worthwhile to return it with indenting and sorted keys.
Note that it is also stored as JSON, but in that cases the JSON is just serialization.
This API json.dumps is effectively a rendering, and the spacing and ordering don't affect
meaning (and make order of keys defined, which is nice), so it doesn't feel too dirty to me.
An alternative would be to JSON.stringify(JSON.parse(data), null, 4) in the UI, but I don't know
how to do that, and at the moment the UI treatment of that data is just as text with the validation
happening on the server, so that'd be new territory.

Test Plan: Unit

Reviewers: vishal

Reviewed By: vishal

Subscribers: changesbot, wwu

Differential Revision: https://tails.corp.dropbox.com/D95663