Add log_processing job listener to report on classifier results for failed job logs
Summary:
This adds the first use of our classification code to track how it does when running against
failure logs. No consequences other than monitoring at the moment, but it'll give us some confidence
in our rules.
Rules are loaded for each log, but that seems insignificant relative to other costs and allows us to
iterate on rules quickly if we need to. Arguable.

Test Plan: Unit

Reviewers: josiah

Reviewed By: josiah

Subscribers: changesbot, wwu, jukka

Differential Revision: https://tails.corp.dropbox.com/D95667