Make trying to resend a collection mail a warning instead of an assertion
Summary:
Tasks are expected to be re-run occasionally if they fail to ack for some reason,
so issuing a warning and terminating successfully seems more appropriate than asserting
something we don't expect to always be true.
We've seen this assertion fail ~1k times in the past couple weeks, and that may be indicative
of other issues, but we can track that with the warning.

Test Plan: None

Reviewers: kajic

Reviewed By: kajic

Subscribers: changesbot, wwu

Differential Revision: https://tails.corp.dropbox.com/D99573