Try to track dropped Jenkins queue jobs more reliably
Summary:
Jobs dropped from the Jenkins queue may be one of the causes of the issue in T31130,
but we can't trust `warning` to show up in Sentry. This promotes it to `exception`, which
we don't have reason to believe is unreliable. Plus, it is an exception, and not a benign one.
Additionally, adds a counter to not need to rely on Sentry exclusively.

Test Plan: None

Reviewers: jukka, vishal

Reviewed By: vishal

Subscribers: changesbot, wwu

Differential Revision: https://tails.corp.dropbox.com/D99103