Make hg VCS raise UnknownRevision as documented
Summary:
We have code that uses `UnknownRevision` as a signal to retry, but hg repos
don't currently benefit from it because they never raise `UnknownRevision`.
This makes `MercurialVcs` raise `UnknownRevision` when hg provides an error
that indicates it is appropriate.
The implementation mirrors the one in `GitVcs`.

Test Plan: Unit

Reviewers: josiah

Reviewed By: josiah

Subscribers: changesbot, wwu

Differential Revision: https://tails.corp.dropbox.com/D100707