Serialize RevisionResult id as a "sha" field too.
Summary:
Revision object shas are serialized as both "id" and "sha", and js uses
sha for a lot of things. The lack of a sha field leads to funny things
like not showing a "#Start" button, and, if there are more than one
RevisionResult objects serialized, the second one overwriting the first
one because we compare by sha.

Reviewers: vishal

Reviewed By: vishal

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D102372