Report failure reasons in build analytics
Summary:
Failure reasons are a great way to track the kinds of failures we see over time, so
we should report them to analytics. It'd be slightly more correct to report them
on a per-jobstep level, but we don't yet have analytics for jobsteps, and
summarizing at the build level isn't incorrect and is useful.

Test Plan: Unit

Reviewers: angie, vishal

Reviewed By: vishal

Subscribers: mkedia, changesbot, wwu

Differential Revision: https://tails.corp.dropbox.com/D104150