Validate the number of steps in the expanded phase of JenkinsTestCollectorStep.
Summary:
This adds a general finish(job)/finish_phase(phase) hook mechanism,
which gives BuildStep implementations a chance to run any extra checks
they might have before marking a Job or JobPhase as passed.
JenkinsTestCollectorStep then uses this to confirm that the proper
number of steps are created for the sharded second phase.

Reviewers: anupc, kylec

Reviewed By: kylec

Subscribers: mkedia, wwu, changesbot

Differential Revision: https://tails.corp.dropbox.com/D102386