Order test-history page according to commit date rather than build date
Summary:
This makes sense and would have made me personally ping the wrong people at least 3 fewer times, nevermind everyone else (e.g. daily pushers) who have the same problem.

We join on `Revision`, and use the `date_committed` which appears to be what we want according to http://stackoverflow.com/questions/11856983/why-git-authordate-is-different-from-commitdate, although it seems the author date would also work since we use `arc` to squash/rebase when landing

Adding a bunch of you to this review, but anyone can stamp it ^_^

Test Plan:
Tested manually:

{F172259}

- The bottom-right is the listing with this change
- The top-right is the listing without this change
- The bottom left is the actual git log.

As you can see, the listing without this change is jumbled, presumably because changes_dev imports things in a arbirary order, while the listing with this change matches the git log precisely.

Reviewers: kylec, josiah, vishal, wwu

Reviewed By: wwu

Subscribers: changesbot, mkedia, wwu

Differential Revision: https://tails.corp.dropbox.com/D106050