Increase frequency of repo updates
Summary:
We rely on repo checks to find new commits in order to trigger
commit builds and use as the basis of diff builds.
The 99th percentile time taken for updating repos is typically less
than 5s, and we regularly see diffs submitted based on revisions we
don't have knowledge of. We can afford to sync repos more often,
and we'll drop fewer diffs and start commit builds faster if we do.

It's worth noting that as far as I can tell, `sync_repo` should
run continuously, but based on what I can see, we sync repos pretty
much exactly ever 5 minutes currently and the `TrackedTask`s responsible
appear finished and passed, which is inconsistent with syncing continuously.
If we understand and resolve that, this change will be inconsequential.

Updates T39008

Test Plan: None

Reviewers: cf

Reviewed By: cf

Subscribers: jukka, changesbot, mkedia, wwu

Differential Revision: https://tails.corp.dropbox.com/D106001