Kill SystemOption, BuildPhase in python code
Summary:
They're not used, so lets get rid of them.

I left `change`/`Change` in place because it's used in one script I'm not clear if we still want that or not.

`itemsequence` still is used by some in-database function so we're leaving it for now

Test Plan:

Tested by manually clicking around, things seem to still work in my dev VM after this removal.

Recursively grepping up the call stack for each of these models did not find any usage except for the `systemoption` API call, of which I could find no callers

Reviewers: mkedia, josiah

Reviewed By: josiah

Subscribers: haoyi, changesbot, mkedia, wwu

Differential Revision: https://tails.corp.dropbox.com/D106632