Use extras.data for uncategorized logs to avoid warningspam in Sentry
Summary:
When all of the warning strings are unique, they aren't grouped in any way, which makes them
incredibly noisy in Sentry. This should resolve that.
We could also use '%s' formatting in `logger.warning` to get the same effect, but this
seems more usefully structured and perhaps less brittle.

Test Plan: Existing unit, with slight tweak to expect `extra`.

Reviewers: anupc

Reviewed By: anupc

Subscribers: changesbot, mkedia, wwu

Differential Revision: https://tails.corp.dropbox.com/D107615