Discard attempted TrackedTasks that don't appear to be in the DB
Summary:
If Celery gives us a TrackedTask we don't remember creating, drop it quietly but report an error.
Though some of the other parts of this class appear to try to work when the Task isn't in the DB,
that doesn't seem to be ever what we want.

Fixes T40944

Test Plan: None yet.

Reviewers: jukka

Reviewed By: jukka

Subscribers: changesbot, mkedia, wwu

Maniphest Tasks: T40944

Differential Revision: https://tails.corp.dropbox.com/D107493