Migrate test scripts from Changes config to ci/
Summary:
Our Changes install should know as little as possible, and moving the configs
into the repo allows us to version them appropriately with the code.
It also allows us to iterate on them without continually tweaking the global
config.
I've intentionally changed nothing so that this is pure relocation.

Test Plan: None

Reviewers: josiah

Reviewed By: josiah

Subscribers: changesbot, mkedia, wwu

Differential Revision: https://tails.corp.dropbox.com/D108164