Place virtualenv outside of the repo in ci/.
Summary:
By moving it outside of the repo, it becomes possible for us to reuse the virtualenv contents
from a snapshot instead of having it be removed by `git clean` every time.
At the moment, we overwrite the virtualenv every time, so this shouldn't have any effect at the moment.
I'm not too worried about storing the env in homedir because the ci script exist for the specific purpose
of being run on Changes where we happily assume the machine is ours to play with.
If we do reuse the virtualenv from a snapshot, my measurements indicate that we'd save about 1min per
Changes build, around 20%.

This also puts timing on the setup step so we can track; currently it takes about 2:17 to do the setup.

Test Plan: Changes

Reviewers: haoyi

Reviewed By: haoyi

Subscribers: changesbot, mkedia, wwu

Differential Revision: https://tails.corp.dropbox.com/D108604