Stop poller before switching pages
Summary:
There's a bug where we poll after we've made the request for a new page but
before the results have come back. This causes us to override the new page
with incorrect rows.

Prevent this from happening

Test Plan:
console.log() to make sure the poller stopped and started when it was
supposed to.

Reviewers: josiah, haoyi

Reviewed By: haoyi

Subscribers: changesbot, mkedia

Maniphest Tasks: T33710

Differential Revision: https://tails.corp.dropbox.com/D108716