Make `last_failure_output` take priority over stdout or stderr
Summary:
Currently, if a test has all of `stdout`, `stderr` and `last_failure_output`, we pick the "first" one, which is pretty arbitrary and always one of `stdout`/`stderr`. This can be see in test cases like

https://changes.build.itc.dropbox.com/projects/servergit/builds/b368238422d44abbb49e3fa0fd7c798a/jobs/56456226ad464d959b9080322209f2bc/tests/16a0e3a48b0b4cdea6d630b09f7b2bdc/

Where it's showing the stdout, whereas the `stderr` and `last_failure_output` present in the artifact:

https://jenkins-server-runner-c.build.itc.dropbox.com/job/generic/167589/artifact/artifacts/junit.xml

(search for `last_expiring_team_sub`)

get ignored. This diff makes it such that when such a conflict exists, pick the `last_failure_output` since it's the most useful to us to debug flakiness in the test. The stdout/stderr tend not to be very useful

Test Plan: Will test in changes dev

Reviewers: tiago

Reviewed By: tiago

Subscribers: changesbot, mkedia, wwu

Differential Revision: https://tails.corp.dropbox.com/D111346