Fix flaky tests graph to filter by project
Summary: The big graph on top of Flaky Tests page was not filtering by project. This is to make it query only flaky tests of a specific project.

Test Plan: Tested locally by loading the flaky tests page for two different projects and making sure that the graph is not the same.

Reviewers: haoyi

Reviewed By: haoyi

Subscribers: mkedia

Differential Revision: https://tails.corp.dropbox.com/D111779