Fix Previous and Next pagination buttons crashing Changes when no previous/next page exists
Summary:
Currently, if you click "Previous" in the first page of any paginated list in Changes (the commits list, for example), you get a JavaScript type error and the UI shows a spinner forever.

This is to fix this behavior by not doing anything when there's no previous/next pages.

Test Plan: Tested by clicking "Previous" in the first page of a paginated list and making sure that nothing happens.

Reviewers: haoyi

Reviewed By: haoyi

Subscribers: mkedia

Differential Revision: https://tails.corp.dropbox.com/D112377