Fix jobstep-allocate crash interaction.
Summary:
For some reason, the mesos schedule doesn't always post resources.mem
and resources.cpu to jobstep_allocate. This data comes from mesos
and it makes more sense to default this on the server
end than leaving it to the scheduler. This fixes the crash on the
individual request which also fixes the BadStatusLine that
the scheduler receives.

Test Plan:
Started changes client with dev_vm and the crash was removed. Change
has, by definition, no effect on behavior in cases where it would not
crash due to the key not being present.

Reviewers: jukka

Reviewed By: jukka

Subscribers: changesbot, mkedia

Differential Revision: https://tails.corp.dropbox.com/D112245