RFC: Add client performance logging
Summary:
Every time we load a page on changes or transition to a new page,
measure the time taken and log it. Also log the amount of time taken by each
API call.

This should enable several useful things:
  a) identify slow api calls
  b) see whether our ajax requests have high variability, e.g. it just takes a
  single api request to be slow (for whatever reason) in order to slow down the
  entire page. Also lets us see if we're sending off so many api calls that some
  are sent later than they need to be
  c) get actual perf numbers on each of our pages

One assumption is that the initial time to show a blank page with load
indicator is actually fairly fast compared to the rest of the page. If we
wanted complete numbers we could look into NavigationTiming, but I've heard
mixed things about how accurate/useful it is. I think this logging does a good
job of capturing everything that's making changes slow.

One caveat: there's still some time taken by the js to render even after all
of this fetching, for example the main project page with a list of builds.
I couldn't find an easy way to include that in the performance metrics (no
angular events), so I'll settle for what we've got.

Test Plan: Added python print statements to the stats reporter, made sure everything looked correct

Reviewers: josiah, haoyi

Reviewed By: haoyi

Subscribers: ptyl, tabbott, vishal, #developer_tools, changesbot, mkedia

Differential Revision: https://tails.corp.dropbox.com/D109270