Add shard_build_type option to test_collector.
Summary:
In many cases it may make sense for sharded builds to operate differently
between the collection phase and each individual shard.

For test_collect, the 'build_type' specifies what to use for the collection
phase; for simple or even static collection phases this probably doesn't need
to be 'lxc' and can either use 'legacy' or 'basic' as long as no provisioning
is necessary.

The 'shard_build_type' specifies how to run each individual shard, and so
eventually we will probably want these running on 'lxc'.

Because the builder for these phases is different, add a kwargs override
in get_builder so that the build plan can do things like:

> self.get_builder(build_type=self.shard_build_type)

and is flexible to any of the other arguments to the builder's constructor.

Test Plan:
Tested locally on changes_dev. Mainly affects build_type-related things
only.

Reviewers: jukka

Reviewed By: jukka

Subscribers: changesbot, vishal, mkedia

Maniphest Tasks: T43972, T43973

Differential Revision: https://tails.corp.dropbox.com/D113268