Avoid creating duplicate jobsteps if task re-tries
Summary: Previously, an exception that interrupted a Changes sync_artifact task as it was creating Jenkins shards would cause the task to re-run and create a second, duplicate set of steps and shards. This diff is more careful, and first tries to fetch existing steps before deciding that none exist and that they must be created.

Test Plan: Add a unit test that successfully forces a double shard build when run against the current code, and then watch it succeed when run against this improved version of the code.

Reviewers: josiah

Reviewed By: josiah

Subscribers: changesbot, mkedia

Projects: #changes, #changes-stability

Differential Revision: https://tails.corp.dropbox.com/D117747