[1/7] Models for garbage collection.
Summary:
Adds the cached snapshot image model, which represents a
garbage-collection-managed snapshot image that may or may
not have an expiration. If expiration date is NULL, then this
means that there is currently no expiration date set.

We cascade on delete so that images which are deleted are automatically
removed from the garbage collection system.

We also add an index to the expiration date because we do comparison
based selects on it.

Finally, add the new model to the testutils fixtures.

Pre-emptive defense for my decision to use a new table for this rather
than add fields to the snapshot_image table: It simplifies the logic because
expiration date field NULL meaning has a meaning already, so adding this
as a field to the main table would:
  - 1: Pollute the main table with Jenkins-specific garbage collection things
  - 2: Need an extra field to label whether it is cached or not

So having a separate table makes more sense here, and handling filtering by joins
is efficient.

Test Plan: Successfully ran the migration and successfully downgraded as well.

Reviewers: armooo

Subscribers: mkedia, vishal

Differential Revision: https://tails.corp.dropbox.com/D119114