[5/7] Add logic to clear gc-database.
Summary:
Expired rows aren't really needed. Note that we aren't using this from
any api endpoint at the moment because I'm not sure where the best place
to call this is.

This commit is optional; it wasn't much work to add this feature and I'm
not particularly attached to it if we don't care about cleaning up this
part of the database.

Test Plan: Unit tests pass. This logic is also not used anywhere yet.

Reviewers: armooo

Subscribers: mkedia, vishal

Differential Revision: https://tails.corp.dropbox.com/D119130