[7/7] API endpoint for caching snapshots.
Summary:
Add an api endpoing that allows the snapshot master to cache a new
snapshot and retrieve information about what clusters to immediately
sync to guarantee that the snapshot is on relevant clusters.

This is the most important endpoint as it is here where we actually
tell the slave to download a new snapshot.

I am not particularly attached to the routings I used for these
two api endpoints. I had a hard time choosing because the interface
is not particularly RESTFUL.

Test Plan:
Tests pass, will do integration tests with this locally but it seems
to work. Code for this entire line of changes is low-risk as it doesn't
interfere with any existing infrastracture.

Reviewers: armooo

Subscribers: mkedia, vishal

Differential Revision: https://tails.corp.dropbox.com/D119133