Add snapshot build finish listener.
Summary:
Right now in the client we don't respond with a snapshot status if the build
fails. This makes sense as the server needs to be able to handle timeouts
and infastracture failures where the status never gets sent. An easy way to
handle this is to use a build finished listener, and for snapshot builds
ensure that the snapshot failed if the build itself failed.

If for some reason the snapshot is active instead of pending, we mark it
as invalidated as active -> failed shouldn't be used as a state transition.

Test Plan: Tested locally on changes_dev + tests pass.

Reviewers: armooo

Reviewed By: armooo

Subscribers: josiah, changesbot, mkedia, jukka, vishal

Maniphest Tasks: T44638

Differential Revision: https://tails.corp.dropbox.com/D119246