Pass in artifact-server as config parameter.
Summary:
We shouldn't use this as a constant. This will allow us to have
a staging version of the artifacts server later and remove
build.itc references from changes-client.

Warning: This depends on a deploy of changes-client that I'm working
on. I'll only land this once that is ready.

Test Plan:
The unit test wasn't passing until I filted out non-None parameters,
so there is verification that this doesn't actually affect parameters
unless config specifies this option.

Reviewers: armooo, paulruan

Reviewed By: paulruan

Subscribers: changesbot, mkedia, jukka, vishal

Differential Revision: https://tails.corp.dropbox.com/D120186