Detect duplicate tests before saving to the db
Summary:
Changes has been relying on a database uniqueness constraint to prevent two results for the same test from being saved for a particular job. This causes two problems. First, the `IntegrityError` raised by the database does not indicate which test was duplicated, so its name could not be reported to the user. Second, the failure of the transaction prevents any test results from being saved to the database, leaving the developer to trawl through the logs to learn what happened.

The solution is to stop relying on the uniqueness constraint to find duplicate tests for us. It is too blunt an instrument.

Instead of letting a duplicate test be considered a major failure of its build, that stops all other test results from being reported, this diff marks only the duplicate test itself as a failure (and removes all but one of the duplicates, to not upset the database). This leaves all other test results alone. This should prevent the database constraint from ever again being triggered, while returning a rich build report to the developer.

Test Plan: I will deploy this patch to staging, trigger a build that deliberately runs several tests twice, and report on the result.

Subscribers: mkedia

Differential Revision: https://tails.corp.dropbox.com/D120487