New commits page UI
Summary:
One of my major goals was to have everything on one page, rather than
the current situation where you have to continually click in to other pages.

I also added a sidebar: the sidebar gives you an overview of what's
happening, as well as other useful data/commands. Its easy to click on a
sidebar link to navigate to the build/job you're interested in.

This page should be in reasonable enough shape to get feedback from others

Note: this is kind of a kitchen sink diff...there's unrelated logic and fixes.
We'll do a pass through the code after we've shipped v1.

Test Plan: Tried it on a bunch of commits

Reviewers: haoyi

Reviewed By: haoyi

Subscribers: changesbot, mkedia

Differential Revision: https://tails.corp.dropbox.com/D120525