Add support for dependent snapshot plans.
Summary:
Adds the ability to configure a plan to use the snapshot image
that another plan built. This is useful if many plans of the same
project are able to use the same snapshot.

Test Plan:
Unit tests/integration tests here should verify that the functionality
we need is present.

Reviewers: armooo

Reviewed By: armooo

Subscribers: changesbot, mkedia, vishal

Differential Revision: https://tails.corp.dropbox.com/D121303