Don't require tests to be reported for snapshots.
Summary:
Snapshot builds will never report tests, but some plans that
have snapshots may want to require test reporting. Currently
these cannot work together, which is unacceptable. The outcome
can be seen by attempting to create a snapshot for such a plan
and seeing that the snapshot build itself seems to pass according
to output, and is marked as active, but then then is marked as
invalidated by a build hook that cleans up weird things like this.

We previously discussed the problem of pending -> active -> invalidated
providing a race condition - but I maintain that any scenario that
allows this sequence to happen at the moment is a bug - as no snapshot
should be marked as active and then invalidated by the build hook.
However, the invalidated sequence does allow us to catch bugs like this.

Test Plan: Unit test verifies the new behavior.

Reviewers: jukka, armooo, kylec

Reviewed By: kylec

Subscribers: changesbot, vishal, mkedia

Maniphest Tasks: T49371

Differential Revision: https://tails.corp.dropbox.com/D122286