[cq] API Endpoint for commit-queue to ensure that builds are created for a commit build
Summary:
There is an existing API for creating commit builds. However, that API doesnt:
1. respect whitelists for commit builds
2. check if builds are already created for this commit.

What we ensure is that we return exactly one build per project. If no builds have been created for a project, we create one. If multiple builds have been created for a project, we return the latest one.

Therefore, we added a new endpoint under /builds/revision/ensure/

Also fixed a bug in test_build_index.py that was not caught because a try...catch is too general.

Test Plan: unit tests

Reviewers: kylec, josiah

Reviewed By: josiah

Subscribers: changesbot, mkedia, #commit_queue

Differential Revision: https://tails.corp.dropbox.com/D121742