Optimize serialize by handling primitive lists and dicts better
Summary:
This codifies the concept of a 'passthrough' type, for which serialization is a no-op and
ensures that any list or dict of only passthrough types given to `serialize` is returned unmodified.
This means most leaf-level dicts and lists are very cheap, and the work required for non-passthrough
dicts is lessened, as the keys/values lists are barely touched; most (all?) dicts are exclusively string keys,
so benefits extend well beyond the leaves.

On my machine, the time taken to (re)process the JSON result for the project list is reduced by ~65%;
the real-life improve is  unlikely to be that good, but we should no longer be calling `get_serializer` once
per dict and should be calling `serialize` one fewer time per dict key, so I still anticipate double-digit
CPU savings.

Test Plan: Existing unit, more if requested.

Reviewers: mkedia

Reviewed By: mkedia

Subscribers: changesbot, wwu, mkedia

Differential Revision: https://tails.corp.dropbox.com/D122730