Use official result aggregation to combine results
Summary: Because it turns out that `max()` cannot be applied to test `Result` values, we should use the official `aggregate_result()` function to help de-duplicate tests for which several different values have been returned in a particular xUnit batch of results (as when py.test lists a test as both having failed and then errored during teardown).

Test Plan: A new unit test detects that we combine quarantined results correctly.

Reviewers: paulruan, tiago

Reviewed By: tiago

Subscribers: changesbot, mkedia

Differential Revision: https://tails.corp.dropbox.com/D124516