Add snapshot_script parameter to generic jenkins.
Summary:
This is useful if there is something that is only useful for the snapshot
phase (use case: creating a cache of data obtained from static analysis)
but doesn't need to be part of the actual shards.

As an example of why we need this at the moment, @fortytwo needs a cached
static analysis blob that is used in the *collection* phase rather than
the shard phase.

Test Plan: New test passes

Reviewers: kylec, jukka

Reviewed By: jukka

Subscribers: changesbot, mkedia, vishal, fortytwo

Differential Revision: https://tails.corp.dropbox.com/D124464