Avoid scanning logsource filtering on job id.
Summary:
This commit takes advantage of the project id index on logsources
which helps significantly when querying for logsources for projects with few of them.
This commit also takes advantage of the fact that jobstep has an index on job id
so it's better to filter on that instead when joining with jobstep.

Test Plan: Pass tests.

Reviewers: mkedia, kylec, naphat

Reviewed By: kylec, naphat

Subscribers: changesbot, mkedia, josiah

Differential Revision: https://tails.corp.dropbox.com/D125623