[cq] updated build_index
Summary:
This diff gets rid of build_revision_ensure API because we decided we want to create a new build everytime, at which point it is easier to consolidate into build_index. This includes a change to build_index behavior:

On specifying a commit build, if the repo (instead of a project) were specified, then we apply project whitelisting and commit-trigger filtering first.

Test Plan: manual + unit tests

Reviewers: kylec, josiah

Reviewed By: josiah

Subscribers: changesbot, mkedia, #commit_queue

Differential Revision: https://tails.corp.dropbox.com/D124798