Don't terminate Jenkins master selection when we fail to contact one
Summary:
When one master of a set goes down, we shouldn't terminate master choice if we
can find another valid master. This changes the Jenkins master selection to only
fail if no master could be chosen.
The previous behavior relied on retries, which may be better in some cases, but
is ugly. Better that we treat master failure as expected.

Test Plan: Ugly unit.

Reviewers: paulruan

Reviewed By: paulruan

Subscribers: changesbot, wwu, mkedia

Differential Revision: https://tails.corp.dropbox.com/D125945