[cq] extra unit tests + comments to build_index
Summary:
This diff attempts to bring some sanity to BuildIndexAPIView by adding explicit comments about what each parameter is doing and what is going on during the API call.

This also includes more unit tests to better capture the behaviors of BuildIndexAPIView that we care about. This will make maintaining compatibility easier.

This diff contains no behavior change, except for the minor case where we now return 400 instead of erroring out 500 when a revision cannot be found in the repo but is already in the database somehow.

Test Plan: ran tests

Reviewers: josiah, paulruan

Reviewed By: paulruan

Subscribers: changesbot, mkedia

Differential Revision: https://tails.corp.dropbox.com/D126062