Second fix for project_test_history page
Summary:
The old fix did not account for the fact that asking for "tests for last 100 commits" could have <100 results, tricking the pagination into thinking that there were no more items to page even though there were. Making the original semantics work is basically impossible given the current state of our data model and pagination logic.

This diff changes the semantics of the test history page from "last 100 test runs, in commit order" to "test runs for last 100 commits" at mkedia's suggestion. This simplifies the page significantly, and defines the behavior in a way that (hopefully) is actually implementable.

As a trade-off, now the test history page may have less than 100 items, if some tests in that range did not run.

As part of the implementation, I made the pagination logic robust enough to take `None`s in place of items. These `None`s would be counted as part of the pagination count w.r.t. determining whether or not there was a next page, but would get removed from the list before it is handed over to the AngularJS page.

Test Plan:
Tested locally, looks like this:

https://www.dropbox.com/s/malejbcfd7jzxw3/Screenshot%202015-07-29%2011.59.26.png?dl=0

Although the page is not full, the `Next` button still works.

Also updated unit tests to verify that this behavior works as intended

Reviewers: mkedia

Reviewed By: mkedia

Subscribers: changesbot, wwu, mkedia

Differential Revision: https://tails.corp.dropbox.com/D126968