Fix typo that stops systemstats polling and lower frequency.
Summary:
We discovered that the queries made are expensive given the current indices.
If we want to keep these stats, on the page, we should fix the queries in the long run but for now,
poll less.
Also, it seems like right now, we keep polling systemstats even if we leave the
admin page, possibly because of the typo.

Test Plan: Test on staging; verify we no longer keep making systemstats requests.

Reviewers: mkedia, kylec

Reviewed By: mkedia, kylec

Subscribers: changesbot, mkedia

Differential Revision: https://tails.corp.dropbox.com/D127492