Updated the builds page
Summary:
The old page gets messy when we have a bunch of builds. I decided to
change a bunch of things
- Simpler, cleaner sidebar
- Made the data fetching code much simpler
- Made the build info page more detailed

If the previous model was a chronological listing of everything that happened
in a build, with a ToC on the side, the new model is more like a task manager
or email client

Note - I'll split this into multiple files in a followup diff.

Test Plan: hit new page

Reviewers: haoyi, anupc

Reviewed By: haoyi

Subscribers: mkedia

Differential Revision: https://tails.corp.dropbox.com/D128271