updated build finished notifier code to take a list of tuple of url and path to CA cert
Summary: This is needed because the requests library comes with its own distribution of root CA certificates, and in order to pass in a custom one (for any *.corp.dropbox.com domains), we need to pass in a special argument.

Test Plan:
unit tests

tested the same `requests` function on a virtualenv on the Changes machine.

Reviewers: josiah

Reviewed By: josiah

Subscribers: changesbot, mkedia, #commit_queue

Differential Revision: https://tails.corp.dropbox.com/D129131