Batch sql queries in the serializer
Summary:
In the status quo, there's basically one place where we need to run
logic to batch multiple data fetches together: when we have a list of objects
with serializers. I wrote a function that will greedily look at the items in
each of these serialized dictionaries, and do a kind of parallel depth-first
traversal, e.g.

The function is passed a list of serialized builds (dicts). It finds the first
key/value pair that needs serialization, e.g. 'source': Source. It serializes
all of them together, resulting in a list of serialized sources. We can
recursively run ourselves on that list to serialize things like the revision
child of source in parallel (which does data fetching.) Now we modify our list
of serialized builds, replacing the source attributes with the serialized
sources we've just run. And return the list of serialized builds.

Quips -
* And people say interview questions have no relation to actual coding
* @kylec, I put you as a reviewer. This is the best welcome back gift I could
  think of :P

Test Plan: arc test. Also, `/api/0/projects/?__trace__=1` is batched!

Reviewers: haoyi, kylec

Reviewed By: kylec

Subscribers: changesbot, mkedia, kylec, jhance, josiah

Differential Revision: https://tails.corp.dropbox.com/D122261