fixed a bug in DiffParser that doesn't handle files that don't end with newlines properly
Summary:
In the unified diff format, if a file does not end with a newline character, an extra line is printed:

\ No new line at end of file

This rarely happens, and even if it does, we only use it to compute coverage information, which is why we never had problems with it. However, we do want to handle it properly if we are to reconstruct a diff from a parsed diff.

Test Plan: unit tests

Reviewers: paulruan, josiah

Reviewed By: paulruan

Subscribers: mkedia

Differential Revision: https://tails.corp.dropbox.com/D130760